fix: Rename tmpl func to distinguish from a normal one with the same name
Though the normal function is declared behind the template one, there still is possibility that the template one can see the normal function, where the bug fixed in this commit is, the tmpl func makes a wrong call to the normal one instead of itself. It is because the instantiation of template function always delay.